requirements.txt
absl-py==0.13.0
aiohttp @ file:///home/conda/feedstock_root/build_artifacts/aiohttp_1623682234273/work
anyio==3.3.1
asgiref==3.4.1
astroid @ file:///tmp/build/80754af9/astroid_1628063140030/work
astunparse==1.6.3
async-timeout==3.0.1
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1620387926260/work
autopep8 @ file:///home/conda/feedstock_root/build_artifacts/autopep8_1619771482781/work
cachetools==4.2.2
certifi==2021.10.8
chardet @ file:///home/conda/feedstock_root/build_artifacts/chardet_1610093492116/work
charset-normalizer==2.0.6
clang==5.0
click==8.0.1
cmake==3.21.2
deepface==0.0.68
discord.py==1.7.3
dlib==19.22.1
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.0.12
Flask==2.0.1
flatbuffers==1.12
gast==0.4.0
gdown==3.13.1
gevent==21.8.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.1
grpcio==1.40.0
h11==0.12.0
h5py==3.1.0
httpcore==0.13.7
httptools==0.2.0
httpx==1.0.0b0
idna==3.2
iniconfig==1.1.1
isort @ file:///tmp/build/80754af9/isort_1602603989581/work
itsdangerous==2.0.1
Jinja2==3.0.1
keras==2.6.0
Keras-Preprocessing==1.1.2
lazy-object-proxy @ file:///tmp/build/80754af9/lazy-object-proxy_1616529027849/work
Markdown==3.3.4
MarkupSafe==2.0.1
mccabe==0.6.1
mtcnn==0.1.1
multidict @ file:///home/conda/feedstock_root/build_artifacts/multidict_1610318999200/work
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.3.56
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.3
Pillow==8.3.2
pluggy==1.0.0
protobuf==3.18.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle @ file:///tmp/build/80754af9/pycodestyle_1615748559966/work
pydantic==1.8.2
pylint @ file:///tmp/build/80754af9/pylint_1627536788603/work
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
retina-face==0.0.5
rfc3986==1.5.0
rsa==4.7.2
six==1.15.0
sniffio==1.2.0
starlette==0.14.2
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
tqdm==4.62.3
typing-extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1632313171031/work
urllib3==1.26.6
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.0
Werkzeug==2.0.1
wrapt @ file:///tmp/build/80754af9/wrapt_1607574498026/work
yarl @ file:///home/conda/feedstock_root/build_artifacts/yarl_1625232865577/work
zope.event==4.5.0
zope.interface==5.4.0
